Input feedback update
Program now displays correctly guessed letters, green if in the correct place, yellow if in a different place, like the Wordle game it's based on. Was advised to not implement color coding at first, but realized it would be the most practical way to implement feedback with the logic I intended to use.